Update ci script

+++ ci/build.ps1
@@ -2,14 +2,18 @@
 # refer to: https://docs.github.com/en/actions/learn-github-actions/environment-variables
 if (!($env:GITHUB_ACTIONS -eq 'true')) {
     $cmake_ver="3.25.2"
-    curl "https://github.com/Kitware/CMake/releases/download/v$cmake_ver/cmake-$cmake_ver-windows-x86_64.zip" -o "cmake-$cmake_ver-windows-x86_64.zip"
-    Expand-Archive -Path cmake-$cmake_ver-windows-x86_64.zip -DestinationPath .\
     $cmake_bin = (Resolve-Path .\cmake-$cmake_ver-windows-x86_64\bin).Path
+    if(!(Test-Path $cmake_bin -PathType Container)) {
+        curl "https://github.com/Kitware/CMake/releases/download/v$cmake_ver/cmake-$cmake_ver-windows-x86_64.zip" -o "cmake-$cmake_ver-windows-x86_64.zip"
+        Expand-Archive -Path cmake-$cmake_ver-windows-x86_64.zip -DestinationPath .\
+    }
     $env:Path = "$cmake_bin;$env:Path"
 }
 
 cmake --version
-
+$buildDir=$(if ($null -ne $env:BUILD_ARCH) {"build_$env:BUILD_ARCH"} else {'build'})
+$CONFIG_ALL_OPTIONS=@()
+$CONFIG_ALL_OPTIONS += "-B$buildDir"
 if ($env:GITHUB_ACTIONS -eq 'true') {
     $T_VS_VERSIONS = @{ '2013' = '12'; '2015' = '14'; '2017' = "15"; '2019' = '16'; '2022' = '17' }
 
@@ -18,7 +22,6 @@ if ($env:GITHUB_ACTIONS -eq 'true') {
     # $env:VS_VERSION: '2013', '2015', '2017', '2019', '2022'
     # $env:TOOLSET: optional, v100
     # $env:CXX_STD: optional, 17, 20, 23
-    $CONFIG_ALL_OPTIONS=@()
 
     $archName = $env:BUILD_ARCH
 
@@ -45,14 +48,14 @@ if ($env:GITHUB_ACTIONS -eq 'true') {
     }
 } else {
     # appveyor ci
-    $CONFIG_ALL_OPTIONS = ($env:GEN -split ',')
+    $CONFIG_ALL_OPTIONS += ($env:GEN -split ',')
 }
 
 Write-Output ("CONFIG_ALL_OPTIONS=$CONFIG_ALL_OPTIONS, Count={0}" -f $CONFIG_ALL_OPTIONS.Count)
 
 ######### Configure
-$buildDir=$(if ($null -ne $env:BUILD_ARCH) {"$env:BUILD_ARCH"} else {'build'})
-cmake -B $buildDir $CONFIG_ALL_OPTIONS
+
+cmake $CONFIG_ALL_OPTIONS
 
 ########## Build
